feat: Consolidate notification module into feature
