+++ app/build.gradle.kts
@@ -65,4 +65,5 @@ dependencies {
 
     implementation(projects.feature.auth)
     implementation(projects.feature.home)
+    implementation(projects.feature.notification)
 }
