+++ feature/notification/build.gradle.kts
@@ -1,0 +1,11 @@
+plugins {
+    alias(libs.plugins.hyunjung.cherrydan.android.feature.ui)
+}
+
+android {
+    namespace = "com.hyunjung.feature.notification"
+}
+
+dependencies {
+    implementation(libs.androidx.navigation.compose)
+}
