+++ settings.gradle.kts
@@ -30,12 +30,10 @@ include(":core:domain")
 include(":core:data")
 include(":core:database")
 include(":core:network")
-include(":notification:data")
-include(":notification:domain")
-include(":notification:presentation")
 include(":search:presentation")
 include(":search:domain")
 include(":search:data")
 include(":core:model")
 include(":feature:auth")
 include(":feature:home")
+include(":feature:notification")
